Fixed peerdependencies and made yarn work.
There were many missing peerDependencies, so this installs _most_ of those. A consequence of this is that yarn now works, so I removed the package-lock.json and added the yarn.lock.

However, there are some peerDependencies that are still not satisfied:

    warning " > jest-styled-components@6.3.1" has incorrect peer dependency "styled-components@^2.0.0 || ^3.0.2".
    warning " > gatsby-plugin-eslint@2.0.1" has incorrect peer dependency "eslint@^4.9.0".

These are not as easy to fix, so I'll leave these for now.