Major link update. All external links in contentful are now references to an external link topic. This meant that the linkhandler can be refactored to better suit the majority use-case. LinkHanlder can take an array due to the multi ref preference in contentful if it the reference takes more than one content model.
LinkHandler now accepts mostly just a link object or array and identifies external link by the internal.type. New test factories created for this.

Migrations and seeds written to update all the references in contentful for consistency. Before some were two single reference. Now in contentful it is a multi ref field that accepts max of 1, that can be a variety of content models (including externalLink).