Legal pages footnotes reworked
Now rich text is rendered using @contentful/rich-text-react-renderer this meant that legal footnotes were not working. Optional parameter added to the RichText rendering that is a callback function for text nodes so that a replacement can be made.

Square bracketed text in the richText are replaced by anchor links  that point to an automatically generated list of footnotes. The list of footnotes and the text inside the anchor link (which is the current count of the footnotes) is generated at render time within the textRenderer callback function. Due to tweaks to RichText snapshot styles className re-ordering has occurred.