Rich text fix
By default RTE when linked to a piece of content references every field on the linked content. The referenced content could then have links to other content in its fields which would extremely easily create circular reference. The patch mutates the linked contents field to just slug to prevent this. This wasn't checking deep enough in the rich text content, as when a link was included inside a link inside a paragraph this was not mutated and caused a loop.